Release cell reference on prepareForReuse as a fix for iOS 15+ cell lifecycle changes.
Due to performance optimizations in iOS 15 and above, the table view data source may generate cells in advance using the cellForRow method. Consequently, a cell can be created but might not go through the complete lifecycle callbacks of willDisplayCell and didEndDisplaying if it remains undisplayed. However, in our implementation, we initiate the process of loading the cell image within cellForRow and only cancel the image request within didEndDisplaying. In scenarios where a cell is reused but was never displayed, a race condition may arise since the request continues, potentially leading to loading an incorrect image at an incorrect index path.